# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Hub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Agent {
  id: number;
  name: string;
  slug: string;
  roleDescription: string;
  systemPrompt: string;
  icon: string;
  color: string;
  isActive: boolean;
  createdAt: string;
}

export interface UpdateAgentBody {
  systemPrompt?: string;
  isActive?: boolean;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  agentId: number;
  businessTag: string;
  createdAt: string;
  updatedAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
  agentId: number;
  businessTag: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  agentId: number;
  businessTag: string;
  createdAt: string;
  updatedAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type BrainDocumentType =
  (typeof BrainDocumentType)[keyof typeof BrainDocumentType];

export const BrainDocumentType = {
  pdf: "pdf",
  text: "text",
  url: "url",
} as const;

export type BrainDocumentBusinessTag =
  (typeof BrainDocumentBusinessTag)[keyof typeof BrainDocumentBusinessTag];

export const BrainDocumentBusinessTag = {
  equifind: "equifind",
  home_inspection: "home_inspection",
  general: "general",
} as const;

export interface BrainDocument {
  id: number;
  title: string;
  type: BrainDocumentType;
  content: string;
  businessTag: BrainDocumentBusinessTag;
  createdAt: string;
}

export type CreateBrainDocumentBodyType =
  (typeof CreateBrainDocumentBodyType)[keyof typeof CreateBrainDocumentBodyType];

export const CreateBrainDocumentBodyType = {
  text: "text",
  url: "url",
} as const;

export type CreateBrainDocumentBodyBusinessTag =
  (typeof CreateBrainDocumentBodyBusinessTag)[keyof typeof CreateBrainDocumentBodyBusinessTag];

export const CreateBrainDocumentBodyBusinessTag = {
  equifind: "equifind",
  home_inspection: "home_inspection",
  general: "general",
} as const;

export interface CreateBrainDocumentBody {
  title: string;
  type: CreateBrainDocumentBodyType;
  content?: string;
  url?: string;
  businessTag: CreateBrainDocumentBodyBusinessTag;
}

export type AutomationStatus =
  (typeof AutomationStatus)[keyof typeof AutomationStatus];

export const AutomationStatus = {
  idle: "idle",
  running: "running",
  pending_approval: "pending_approval",
} as const;

export interface Automation {
  id: number;
  name: string;
  agentId: number;
  agentName: string;
  agentIcon: string;
  agentColor: string;
  scheduleCron?: string | null;
  promptTemplate: string;
  lastRanAt?: string | null;
  nextRunAt?: string | null;
  isActive: boolean;
  status: AutomationStatus;
  lastOutput?: string | null;
  createdAt: string;
}

export interface CreateAutomationBody {
  name: string;
  agentId: number;
  scheduleCron?: string;
  promptTemplate: string;
  isActive?: boolean;
}

export interface UpdateAutomationBody {
  name?: string;
  scheduleCron?: string;
  promptTemplate?: string;
  isActive?: boolean;
}

export type AutomationRunStatus =
  (typeof AutomationRunStatus)[keyof typeof AutomationRunStatus];

export const AutomationRunStatus = {
  success: "success",
  failed: "failed",
  pending_approval: "pending_approval",
} as const;

export interface AutomationRun {
  id: number;
  automationId: number;
  automationName: string;
  agentName: string;
  agentIcon: string;
  agentColor: string;
  output?: string | null;
  status: AutomationRunStatus;
  ranAt: string;
}

export type ConnectionPlatform =
  (typeof ConnectionPlatform)[keyof typeof ConnectionPlatform];

export const ConnectionPlatform = {
  linkedin: "linkedin",
  google: "google",
  twitter: "twitter",
  meta: "meta",
  gohighlevel: "gohighlevel",
  email: "email",
} as const;

export type ConnectionAuthType =
  (typeof ConnectionAuthType)[keyof typeof ConnectionAuthType];

export const ConnectionAuthType = {
  oauth: "oauth",
  api_key: "api_key",
} as const;

export interface Connection {
  id: number;
  platform: ConnectionPlatform;
  displayName: string;
  accountLabel?: string | null;
  isConnected: boolean;
  authType: ConnectionAuthType;
  scopes: string[];
  createdAt: string;
  expiresAt?: string | null;
}

export type CreateConnectionBodyPlatform =
  (typeof CreateConnectionBodyPlatform)[keyof typeof CreateConnectionBodyPlatform];

export const CreateConnectionBodyPlatform = {
  gohighlevel: "gohighlevel",
  email: "email",
} as const;

export type CreateConnectionBodyMetadata = { [key: string]: unknown };

export interface CreateConnectionBody {
  platform: CreateConnectionBodyPlatform;
  apiKey: string;
  accountLabel?: string;
  metadata?: CreateConnectionBodyMetadata;
}

export type ConnectionTestResultAccountInfo = { [key: string]: unknown } | null;

export interface ConnectionTestResult {
  success: boolean;
  message: string;
  accountInfo?: ConnectionTestResultAccountInfo;
}

export interface OAuthInitiateResponse {
  url: string;
  state: string;
}

export interface ConnectionPostBody {
  content: string;
  mediaUrls?: string[];
  targetId?: string | null;
}

export interface ConnectionSendMessageBody {
  recipientId: string;
  message: string;
}

export interface ConnectionActionResult {
  success: boolean;
  message: string;
  postId?: string | null;
}

export type ListAnthropicConversationsParams = {
  agentId?: number;
};

export type ListBrainDocumentsParams = {
  businessTag?: ListBrainDocumentsBusinessTag;
};

export type ListBrainDocumentsBusinessTag =
  (typeof ListBrainDocumentsBusinessTag)[keyof typeof ListBrainDocumentsBusinessTag];

export const ListBrainDocumentsBusinessTag = {
  equifind: "equifind",
  home_inspection: "home_inspection",
  general: "general",
} as const;

export type UploadBrainDocumentBodyBusinessTag =
  (typeof UploadBrainDocumentBodyBusinessTag)[keyof typeof UploadBrainDocumentBodyBusinessTag];

export const UploadBrainDocumentBodyBusinessTag = {
  equifind: "equifind",
  home_inspection: "home_inspection",
  general: "general",
} as const;

export type UploadBrainDocumentBody = {
  file: Blob;
  title: string;
  businessTag: UploadBrainDocumentBodyBusinessTag;
};

export type ListAutomationRunsParams = {
  automationId?: number;
  status?: ListAutomationRunsStatus;
};

export type ListAutomationRunsStatus =
  (typeof ListAutomationRunsStatus)[keyof typeof ListAutomationRunsStatus];

export const ListAutomationRunsStatus = {
  success: "success",
  failed: "failed",
  pending_approval: "pending_approval",
} as const;

export type OauthCallbackParams = {
  code?: string;
  state?: string;
  error?: string;
};