# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"

  # ─── Conversations ────────────────────────────────────────────────────────
...
  /brain/documents:
    get:
      operationId: listBrainDocuments
      tags: [brain]
      summary: List all brain documents
      parameters:
        - name: businessTag
          in: query
          required: false
          schema:
            type: string
            enum: [equifind, home_inspection, general]
      responses:
        "200":
          description: List of brain documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrainDocument"
    post:
      operationId: createBrainDocument
      tags: [brain]
      summary: Create a brain document (text or URL)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrainDocumentBody"
      responses:
        "201":
          description: Created document(s)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrainDocument"
  /brain/documents/upload:
    post:
      operationId: uploadBrainDocument
      tags: [brain]
      summary: Upload a PDF brain document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                businessTag:
                  type: string
                  enum: [equifind, home_inspection, general]
              required: [file, title, businessTag]
      responses:
        "201":
          description: Uploaded and processed documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrainDocument"
  /brain/documents/{id}:
    delete:
      operationId: deleteBrainDocument
      tags: [brain]
      summary: Delete a brain document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Automations ─────────────────────────────────────────────────────────
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List all automations
      responses:
        "200":
          description: List of automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create a new automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationBody"
      responses:
        "201":
          description: Created automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}:
    get:
      operationId: getAutomation
      tags: [automations]
      summary: Get automation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    patch:
      operationId: updateAutomation
      tags: [automations]
      summary: Update automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationBody"
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /automations/{id}/run:
    post:
      operationId: runAutomation
      tags: [automations]
      summary: Run automation immediately
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRun"
  /automations/runs:
    get:
      operationId: listAutomationRuns
      tags: [automations]
      summary: List automation runs
      parameters:
        - name: automationId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [success, failed, pending_approval]
      responses:
        "200":
          description: List of automation runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationRun"
  /automations/runs/{id}/approve:
    post:
      operationId: approveAutomationRun
      tags: [automations]
      summary: Approve a pending automation run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRun"
  /automations/runs/{id}/discard:
    post:
      operationId: discardAutomationRun
      tags: [automations]
      summary: Discard a pending automation run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discarded run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRun"

  # ─── Connections ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Agents ───────────────────────────────────────────────────────────
    Agent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        roleDescription:
          type: string
        systemPrompt:
          type: string
        icon:
          type: string
        color:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - roleDescription
        - systemPrompt
        - icon
        - color
        - isActive
        - createdAt
    UpdateAgentBody:
      type: object
      properties:
        systemPrompt:
          type: string
        isActive:
          type: boolean

    # ─── Conversations / Anthropic ────────────────────────────────────────
...
    BrainDocument:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [pdf, text, url]
        content:
          type: string
        businessTag:
          type: string
          enum: [equifind, home_inspection, general]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - content
        - businessTag
        - createdAt
    CreateBrainDocumentBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [text, url]
        content:
          type: string
        url:
          type: string
        businessTag:
          type: string
          enum: [equifind, home_inspection, general]
      required:
        - title
        - type
        - businessTag

    # ─── Automations ─────────────────────────────────────────────────────
    Automation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        agentId:
          type: integer
        agentName:
          type: string
        agentIcon:
          type: string
        agentColor:
          type: string
        scheduleCron:
          type: string
          nullable: true
        promptTemplate:
          type: string
        lastRanAt:
          type: string
          format: date-time
          nullable: true
        nextRunAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        status:
          type: string
          enum: [idle, running, pending_approval]
        lastOutput:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - agentId
        - agentName
        - agentIcon
        - agentColor
        - promptTemplate
        - isActive
        - status
        - createdAt
    CreateAutomationBody:
      type: object
      properties:
        name:
          type: string
        agentId:
          type: integer
        scheduleCron:
          type: string
        promptTemplate:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - agentId
        - promptTemplate
    UpdateAutomationBody:
      type: object
      properties:
        name:
          type: string
        scheduleCron:
          type: string
        promptTemplate:
          type: string
        isActive:
          type: boolean
    AutomationRun:
      type: object
      properties:
        id:
          type: integer
        automationId:
          type: integer
        automationName:
          type: string
        agentName:
          type: string
        agentIcon:
          type: string
        agentColor:
          type: string
        output:
          type: string
          nullable: true
        status:
          type: string
          enum: [success, failed, pending_approval]
        ranAt:
          type: string
          format: date-time
      required:
        - id
        - automationId
        - automationName
        - agentName
        - agentIcon
        - agentColor
        - status
        - ranAt

    # ─── Connections ─────────────────────────────────────────────────────
...