# lib/api-zod/src/generated/types/automation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AutomationStatus } from "./automationStatus";

export interface Automation {
  id: number;
  name: string;
  agentId: number;
  agentName: string;
  agentIcon: string;
  agentColor: string;
  scheduleCron?: string | null;
  promptTemplate: string;
  lastRanAt?: Date | null;
  nextRunAt?: Date | null;
  isActive: boolean;
  status: AutomationStatus;
  lastOutput?: string | null;
  createdAt: Date;
}