# lib/api-zod/src/generated/types/brainDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BrainDocumentBusinessTag } from "./brainDocumentBusinessTag";
import type { BrainDocumentType } from "./brainDocumentType";

export interface BrainDocument {
  id: number;
  title: string;
  type: BrainDocumentType;
  content: string;
  businessTag: BrainDocumentBusinessTag;
  createdAt: Date;
}