# lib/api-zod/src/generated/types/updateAutomationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAutomationBody {
  name?: string;
  scheduleCron?: string;
  promptTemplate?: string;
  isActive?: boolean;
}